refactored math code adapted quadtree for high-precision improved show_quadtree debug mode
